refactor(plugins): extract shared SQL filter builder and numeric helpers, split oversized drivers
