+++ Plugins/RedisDriverPlugin/RedisPluginDriver.swift
@@ -11,19 +11,19 @@ import Foundation
 import OSLog
 import TableProPluginKit
 
-private extension Array where Element == String? {
+extension Array where Element == String? {
     var asCells: [PluginCellValue] { map(PluginCellValue.fromOptional) }
 }
 
-private extension Array where Element == String {
+extension Array where Element == String {
     var asCells: [PluginCellValue] { map(PluginCellValue.text) }
 }
 
-private extension Array where Element == [String?] {
+extension Array where Element == [String?] {
     var asCellRows: [[PluginCellValue]] { map { $0.map(PluginCellValue.fromOptional) } }
 }
 
-private extension Array where Element == [String] {
+extension Array where Element == [String] {
     var asCellRows: [[PluginCellValue]] { map { $0.map(PluginCellValue.text) } }
 }
 
@@ -35,8 +35,8 @@ final class RedisPluginDriver: PluginDatabaseDriver, @unchecked Sendable {
 
     private static let maxScanKeys = PluginRowLimits.emergencyMax
 
-    private var cachedScanPattern: String?
-    private var cachedScanKeys: [String]?
+    var cachedScanPattern: String?
+    var cachedScanKeys: [String]?
 
     var serverVersion: String? {
         redisConnection?.serverVersion()
@@ -627,1069 +627,6 @@ final class RedisPluginDriver: PluginDatabaseDriver, @unchecked Sendable {
         return generator.generateStatements(
             from: changes, insertedRowData: insertedRowData,
             deletedRowIndices: deletedRowIndices, insertedRowIndices: insertedRowIndices
-        )
-    }
-}
-
-// MARK: - Operation Dispatch
-
-private extension RedisPluginDriver {
-    func executeOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .get, .set, .del, .keys, .scan, .type, .ttl, .pttl, .expire, .persist, .rename, .exists:
-            return try await executeKeyOperation(operation, connection: conn, startTime: startTime)
-
-        case .hget, .hset, .hgetall, .hdel:
-            return try await executeHashOperation(operation, connection: conn, startTime: startTime)
-
-        case .lrange, .lpush, .rpush, .llen:
-            return try await executeListOperation(operation, connection: conn, startTime: startTime)
-
-        case .smembers, .sadd, .srem, .scard:
-            return try await executeSetOperation(operation, connection: conn, startTime: startTime)
-
-        case .zrange, .zadd, .zrem, .zcard:
-            return try await executeSortedSetOperation(operation, connection: conn, startTime: startTime)
-
-        case .xrange, .xlen:
-            return try await executeStreamOperation(operation, connection: conn, startTime: startTime)
-
-        case .ping, .info, .dbsize, .flushdb, .select, .configGet, .configSet, .command, .multi, .exec, .discard:
-            return try await executeServerOperation(operation, connection: conn, startTime: startTime)
-        }
-    }
-
-    // MARK: - Key Operations
-
-    func executeKeyOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .get(let key):
-            let result = try await conn.executeCommand(["GET", key])
-            let value = result.stringValue
-            return PluginQueryResult(
-                columns: ["Key", "Value"],
-                columnTypeNames: ["String", "String"],
-                rows: [[key, value].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .set(let key, let value, let options):
-            var args = ["SET", key, value]
-            if let opts = options {
-                if let ex = opts.ex { args += ["EX", String(ex)] }
-                if let px = opts.px { args += ["PX", String(px)] }
-                if let exat = opts.exat { args += ["EXAT", String(exat)] }
-                if let pxat = opts.pxat { args += ["PXAT", String(pxat)] }
-                if opts.nx { args.append("NX") }
-                if opts.xx { args.append("XX") }
-            }
-            _ = try await conn.executeCommand(args)
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .del(let keys):
-            let args = ["DEL"] + keys
-            let result = try await conn.executeCommand(args)
-            let deleted = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["deleted"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(deleted)].asCells],
-                rowsAffected: deleted,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .keys(let pattern):
-            let result = try await conn.executeCommand(["KEYS", pattern])
-            guard let items = result.arrayValue else {
-                return buildEmptyKeyResult(startTime: startTime)
-            }
-            let keys = items.map { redisReplyToString($0) }
-            let capped = Array(keys.prefix(PluginRowLimits.emergencyMax))
-            let keysTruncated = keys.count > PluginRowLimits.emergencyMax
-            return try await buildKeyBrowseResult(
-                keys: capped, connection: conn, startTime: startTime, isTruncated: keysTruncated
-            )
-
-        case .scan(let cursor, let pattern, let count):
-            var args = ["SCAN", String(cursor)]
-            if let p = pattern { args += ["MATCH", p] }
-            if let c = count { args += ["COUNT", String(c)] }
-            let result = try await conn.executeCommand(args)
-            return try await handleScanResult(result, connection: conn, startTime: startTime)
-
-        case .type(let key):
-            let result = try await conn.executeCommand(["TYPE", key])
-            let typeName = result.stringValue ?? "none"
-            return PluginQueryResult(
-                columns: ["Key", "Type"],
-                columnTypeNames: ["String", "String"],
-                rows: [[key, typeName].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .ttl(let key):
-            let result = try await conn.executeCommand(["TTL", key])
-            let ttl = result.intValue ?? -1
-            return PluginQueryResult(
-                columns: ["Key", "TTL"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(ttl)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .pttl(let key):
-            let result = try await conn.executeCommand(["PTTL", key])
-            let pttl = result.intValue ?? -1
-            return PluginQueryResult(
-                columns: ["Key", "PTTL"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(pttl)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .expire(let key, let seconds):
-            let result = try await conn.executeCommand(["EXPIRE", key, String(seconds)])
-            let success = (result.intValue ?? 0) == 1
-            return buildStatusResult(success ? "OK" : "Key not found", startTime: startTime)
-
-        case .persist(let key):
-            let result = try await conn.executeCommand(["PERSIST", key])
-            let success = (result.intValue ?? 0) == 1
-            return buildStatusResult(success ? "OK" : "Key not found or no TTL", startTime: startTime)
-
-        case .rename(let key, let newKey):
-            let reply = try await conn.executeCommand(["RENAME", key, newKey])
-            if case .error(let msg) = reply {
-                throw RedisPluginError(code: 0, message: "RENAME failed: \(msg)")
-            }
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .exists(let keys):
-            let args = ["EXISTS"] + keys
-            let result = try await conn.executeCommand(args)
-            let count = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["exists"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(count)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeKeyOperation")
-        }
-    }
-
-    // MARK: - Hash Operations
-
-    func executeHashOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .hget(let key, let field):
-            let result = try await conn.executeCommand(["HGET", key, field])
-            let value = result.stringValue
-            return PluginQueryResult(
-                columns: ["Field", "Value"],
-                columnTypeNames: ["String", "String"],
-                rows: [[field, value].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .hset(let key, let fieldValues):
-            var args = ["HSET", key]
-            for (field, value) in fieldValues {
-                args += [field, value]
-            }
-            let result = try await conn.executeCommand(args)
-            let added = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["added"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(added)].asCells],
-                rowsAffected: added,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .hgetall(let key):
-            let result = try await conn.executeCommand(["HGETALL", key])
-            return buildHashResult(result, startTime: startTime)
-
-        case .hdel(let key, let fields):
-            let args = ["HDEL", key] + fields
-            let result = try await conn.executeCommand(args)
-            let removed = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["removed"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(removed)].asCells],
-                rowsAffected: removed,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeHashOperation")
-        }
-    }
-
-    // MARK: - List Operations
-
-    func executeListOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .lrange(let key, let start, let stop):
-            let result = try await conn.executeCommand(["LRANGE", key, String(start), String(stop)])
-            return buildListResult(result, startOffset: start, startTime: startTime)
-
-        case .lpush(let key, let values):
-            let args = ["LPUSH", key] + values
-            let result = try await conn.executeCommand(args)
-            let length = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["length"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(length)].asCells],
-                rowsAffected: values.count,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .rpush(let key, let values):
-            let args = ["RPUSH", key] + values
-            let result = try await conn.executeCommand(args)
-            let length = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["length"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(length)].asCells],
-                rowsAffected: values.count,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .llen(let key):
-            let result = try await conn.executeCommand(["LLEN", key])
-            let length = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["Key", "Length"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(length)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeListOperation")
-        }
-    }
-
-    // MARK: - Set Operations
-
-    func executeSetOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .smembers(let key):
-            let result = try await conn.executeCommand(["SMEMBERS", key])
-            return buildSetResult(result, startTime: startTime)
-
-        case .sadd(let key, let members):
-            let args = ["SADD", key] + members
-            let result = try await conn.executeCommand(args)
-            let added = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["added"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(added)].asCells],
-                rowsAffected: added,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .srem(let key, let members):
-            let args = ["SREM", key] + members
-            let result = try await conn.executeCommand(args)
-            let removed = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["removed"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(removed)].asCells],
-                rowsAffected: removed,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .scard(let key):
-            let result = try await conn.executeCommand(["SCARD", key])
-            let count = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["Key", "Cardinality"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(count)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeSetOperation")
-        }
-    }
-
-    // MARK: - Sorted Set Operations
-
-    func executeSortedSetOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .zrange(let key, let start, let stop, let flags):
-            var args = ["ZRANGE", key, start, stop]
-            args += flags
-            let withScores = flags.contains("WITHSCORES")
-            let result = try await conn.executeCommand(args)
-            return buildSortedSetResult(result, withScores: withScores, startTime: startTime)
-
-        case .zadd(let key, let flags, let scoreMembers):
-            var args = ["ZADD", key]
-            args += flags
-            for (score, member) in scoreMembers {
-                args += [String(score), member]
-            }
-            let result = try await conn.executeCommand(args)
-            if flags.contains("INCR") {
-                // INCR mode returns the new score (or nil for NX miss)
-                let scoreStr = result.stringValue ?? "nil"
-                return PluginQueryResult(
-                    columns: ["score"],
-                    columnTypeNames: ["String"],
-                    rows: [[scoreStr].asCells],
-                    rowsAffected: 0,
-                    executionTime: Date().timeIntervalSince(startTime)
-                )
-            }
-            let count = result.intValue ?? 0
-            let columnName = flags.contains("CH") ? "changed" : "added"
-            return PluginQueryResult(
-                columns: [columnName],
-                columnTypeNames: ["Int64"],
-                rows: [[String(count)].asCells],
-                rowsAffected: count,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .zrem(let key, let members):
-            let args = ["ZREM", key] + members
-            let result = try await conn.executeCommand(args)
-            let removed = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["removed"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(removed)].asCells],
-                rowsAffected: removed,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .zcard(let key):
-            let result = try await conn.executeCommand(["ZCARD", key])
-            let count = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["Key", "Cardinality"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(count)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeSortedSetOperation")
-        }
-    }
-
-    // MARK: - Stream Operations
-
-    func executeStreamOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .xrange(let key, let start, let end, let count):
-            var args = ["XRANGE", key, start, end]
-            if let c = count { args += ["COUNT", String(c)] }
-            let result = try await conn.executeCommand(args)
-            return buildStreamResult(result, startTime: startTime)
-
-        case .xlen(let key):
-            let result = try await conn.executeCommand(["XLEN", key])
-            let length = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["Key", "Length"],
-                columnTypeNames: ["String", "Int64"],
-                rows: [[key, String(length)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeStreamOperation")
-        }
-    }
-
-    // MARK: - Server Operations
-
-    func executeServerOperation(
-        _ operation: RedisOperation,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        switch operation {
-        case .ping:
-            _ = try await conn.executeCommand(["PING"])
-            return PluginQueryResult(
-                columns: ["ok"],
-                columnTypeNames: ["Int32"],
-                rows: [["1"].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .info(let section):
-            var args = ["INFO"]
-            if let s = section { args.append(s) }
-            let result = try await conn.executeCommand(args)
-            let infoText = result.stringValue ?? String(describing: result)
-            return PluginQueryResult(
-                columns: ["info"],
-                columnTypeNames: ["String"],
-                rows: [[infoText].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .dbsize:
-            let result = try await conn.executeCommand(["DBSIZE"])
-            let count = result.intValue ?? 0
-            return PluginQueryResult(
-                columns: ["keys"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(count)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .flushdb:
-            _ = try await conn.executeCommand(["FLUSHDB"])
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .select(let database):
-            try await conn.selectDatabase(database)
-            cachedScanPattern = nil
-            cachedScanKeys = nil
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .configGet(let parameter):
-            let result = try await conn.executeCommand(["CONFIG", "GET", parameter])
-            return buildConfigResult(result, startTime: startTime)
-
-        case .configSet(let parameter, let value):
-            _ = try await conn.executeCommand(["CONFIG", "SET", parameter, value])
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .command(let args):
-            let result = try await conn.executeCommand(args)
-            return buildGenericResult(result, startTime: startTime)
-
-        case .multi:
-            _ = try await conn.executeCommand(["MULTI"])
-            return buildStatusResult("OK", startTime: startTime)
-
-        case .exec:
-            let result = try await conn.executeCommand(["EXEC"])
-            return buildGenericResult(result, startTime: startTime)
-
-        case .discard:
-            _ = try await conn.executeCommand(["DISCARD"])
-            return buildStatusResult("OK", startTime: startTime)
-
-        default:
-            throw RedisPluginError(code: 0, message: "Unexpected operation in executeServerOperation")
-        }
-    }
-}
-
-// MARK: - SCAN Helpers
-
-private extension RedisPluginDriver {
-    func scanAllKeys(
-        connection conn: RedisPluginConnection,
-        pattern: String?,
-        maxKeys: Int
-    ) async throws -> [String] {
-        var allKeys: [String] = []
-        var cursor = "0"
-
-        repeat {
-            var args = ["SCAN", cursor]
-            if let p = pattern {
-                args += ["MATCH", p]
-            }
-            args += ["COUNT", "1000"]
-
-            let result = try await conn.executeCommand(args)
-
-            guard case .array(let scanResult) = result,
-                  scanResult.count == 2 else {
-                break
-            }
-
-            let nextCursor: String
-            switch scanResult[0] {
-            case .string(let s): nextCursor = s
-            case .status(let s): nextCursor = s
-            case .data(let d): nextCursor = String(data: d, encoding: .utf8) ?? "0"
-            default: nextCursor = "0"
-            }
-            cursor = nextCursor
-
-            if case .array(let keyReplies) = scanResult[1] {
-                for reply in keyReplies {
-                    switch reply {
-                    case .string(let k): allKeys.append(k)
-                    case .data(let d):
-                        if let k = String(data: d, encoding: .utf8) { allKeys.append(k) }
-                    default: break
-                    }
-                }
-            }
-
-            if allKeys.count >= maxKeys {
-                allKeys = Array(allKeys.prefix(maxKeys))
-                break
-            }
-        } while cursor != "0"
-
-        return allKeys.sorted()
-    }
-
-    func handleScanResult(
-        _ result: RedisReply,
-        connection conn: RedisPluginConnection,
-        startTime: Date
-    ) async throws -> PluginQueryResult {
-        guard case .array(let scanResult) = result,
-              scanResult.count == 2,
-              case .array(let keyReplies) = scanResult[1] else {
-            return buildEmptyKeyResult(startTime: startTime)
-        }
-
-        let keys = keyReplies.compactMap { reply -> String? in
-            if case .string(let k) = reply { return k }
-            if case .data(let d) = reply { return String(data: d, encoding: .utf8) }
-            return nil
-        }
-
-        let capped = Array(keys.prefix(PluginRowLimits.emergencyMax))
-        let keysTruncated = keys.count > PluginRowLimits.emergencyMax
-        return try await buildKeyBrowseResult(
-            keys: capped, connection: conn, startTime: startTime, isTruncated: keysTruncated
-        )
-    }
-}
-
-// MARK: - Result Building
-
-private extension RedisPluginDriver {
-    static let previewLimit = 100
-    static let previewMaxChars = 1_000
-
-    func buildKeyBrowseResult(
-        keys: [String],
-        connection conn: RedisPluginConnection,
-        startTime: Date,
-        isTruncated: Bool = false
-    ) async throws -> PluginQueryResult {
-        guard !keys.isEmpty else {
-            return buildEmptyKeyResult(startTime: startTime)
-        }
-
-        var typeAndTtlCommands: [[String]] = []
-        typeAndTtlCommands.reserveCapacity(keys.count * 2)
-        for key in keys {
-            typeAndTtlCommands.append(["TYPE", key])
-            typeAndTtlCommands.append(["TTL", key])
-        }
-        let typeAndTtlReplies = try await conn.executePipeline(typeAndTtlCommands)
-
-        var typeNames: [String] = []
-        typeNames.reserveCapacity(keys.count)
-        var ttlValues: [Int] = []
-        ttlValues.reserveCapacity(keys.count)
-        for i in 0 ..< keys.count {
-            let typeName = (typeAndTtlReplies[i * 2].stringValue ?? "unknown").uppercased()
-            let ttl = typeAndTtlReplies[i * 2 + 1].intValue ?? -1
-            typeNames.append(typeName)
-            ttlValues.append(ttl)
-        }
-
-        var previewCommands: [[String]] = []
-        previewCommands.reserveCapacity(keys.count)
-        var previewCommandIndices: [Int] = []
-        previewCommandIndices.reserveCapacity(keys.count)
-
-        for (i, key) in keys.enumerated() {
-            let command: [String]? = previewCommandForType(typeNames[i], key: key)
-            if let command {
-                previewCommandIndices.append(previewCommands.count)
-                previewCommands.append(command)
-            } else {
-                previewCommandIndices.append(-1)
-            }
-        }
-
-        var previewReplies: [RedisReply] = []
-        if !previewCommands.isEmpty {
-            previewReplies = try await conn.executePipeline(previewCommands)
-        }
-
-        var rows: [[PluginCellValue]] = []
-        rows.reserveCapacity(keys.count)
-        for (i, key) in keys.enumerated() {
-            let ttlStr = String(ttlValues[i])
-            let pipelineIndex = previewCommandIndices[i]
-            let preview: String?
-            if pipelineIndex >= 0, pipelineIndex < previewReplies.count {
-                preview = formatPreviewReply(
-                    previewReplies[pipelineIndex], type: typeNames[i]
-                )
-            } else {
-                preview = nil
-            }
-            rows.append([key, typeNames[i], ttlStr, preview].asCells)
-        }
-
-        return PluginQueryResult(
-            columns: ["Key", "Type", "TTL", "Value"],
-            columnTypeNames: ["String", "RedisType", "RedisInt", "RedisRaw"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime),
-            isTruncated: isTruncated
-        )
-    }
-
-    func previewCommandForType(_ type: String, key: String) -> [String]? {
-        switch type.lowercased() {
-        case "string":
-            return ["GET", key]
-        case "hash":
-            return ["HSCAN", key, "0", "COUNT", String(Self.previewLimit)]
-        case "list":
-            return ["LRANGE", key, "0", String(Self.previewLimit - 1)]
-        case "set":
-            return ["SSCAN", key, "0", "COUNT", String(Self.previewLimit)]
-        case "zset":
-            return ["ZRANGE", key, "0", String(Self.previewLimit - 1), "WITHSCORES"]
-        case "stream":
-            return ["XREVRANGE", key, "+", "-", "COUNT", "5"]
-        default:
-            return nil
-        }
-    }
-
-    func formatPreviewReply(_ reply: RedisReply, type: String) -> String? {
-        switch type.lowercased() {
-        case "string":
-            return truncatePreview(redisReplyToString(reply))
-
-        case "hash":
-            let array: [RedisReply]
-            if case .array(let scanResult) = reply,
-               scanResult.count == 2,
-               let items = scanResult[1].arrayValue {
-                array = items
-            } else if let items = reply.arrayValue, !items.isEmpty {
-                array = items
-            } else {
-                return "{}"
-            }
-            guard !array.isEmpty else { return "{}" }
-            var pairs: [String] = []
-            var idx = 0
-            while idx + 1 < array.count {
-                let field = redisReplyToString(array[idx])
-                let value = redisReplyToString(array[idx + 1])
-                pairs.append(
-                    "\"\(escapeJsonString(field))\":\"\(escapeJsonString(value))\""
-                )
-                idx += 2
-            }
-            return truncatePreview("{\(pairs.joined(separator: ","))}")
-
-        case "list":
-            guard let items = reply.arrayValue else { return "[]" }
-            let quoted = items.map { "\"\(escapeJsonString(redisReplyToString($0)))\"" }
-            return truncatePreview("[\(quoted.joined(separator: ", "))]")
-
-        case "set":
-            let members: [RedisReply]
-            if case .array(let scanResult) = reply,
-               scanResult.count == 2,
-               let items = scanResult[1].arrayValue {
-                members = items
-            } else if let items = reply.arrayValue {
-                members = items
-            } else {
-                return "[]"
-            }
-            let quoted = members.map { "\"\(escapeJsonString(redisReplyToString($0)))\"" }
-            return truncatePreview("[\(quoted.joined(separator: ", "))]")
-
-        case "zset":
-            // Parse WITHSCORES result: alternating member, score pairs
-            guard let items = reply.arrayValue, !items.isEmpty else { return "[]" }
-            var pairs: [String] = []
-            var i = 0
-            while i + 1 < items.count {
-                pairs.append("\(redisReplyToString(items[i])):\(redisReplyToString(items[i + 1]))")
-                i += 2
-            }
-            return truncatePreview(pairs.joined(separator: ", "))
-
-        case "stream":
-            // Parse XREVRANGE result: array of [id, [field, value, ...]] entries
-            guard let entries = reply.arrayValue, !entries.isEmpty else {
-                return "(0 entries)"
-            }
-            var entryStrings: [String] = []
-            for entry in entries {
-                guard let parts = entry.arrayValue, parts.count >= 2,
-                      let fields = parts[1].arrayValue else {
-                    continue
-                }
-                let entryId = redisReplyToString(parts[0])
-                var fieldPairs: [String] = []
-                var j = 0
-                while j + 1 < fields.count {
-                    fieldPairs.append("\(redisReplyToString(fields[j]))=\(redisReplyToString(fields[j + 1]))")
-                    j += 2
-                }
-                entryStrings.append("\(entryId): \(fieldPairs.joined(separator: ", "))")
-            }
-            return truncatePreview(entryStrings.joined(separator: "; "))
-
-        default:
-            return nil
-        }
-    }
-
-    func truncatePreview(_ value: String?) -> String? {
-        guard let value else { return nil }
-        let nsValue = value as NSString
-        if nsValue.length > Self.previewMaxChars {
-            return nsValue.substring(to: Self.previewMaxChars) + "..."
-        }
-        return value
-    }
-
-    func escapeJsonString(_ str: String) -> String {
-        var result = ""
-        for scalar in str.unicodeScalars {
-            switch scalar {
-            case "\\": result += "\\\\"
-            case "\"": result += "\\\""
-            case "\n": result += "\\n"
-            case "\r": result += "\\r"
-            case "\t": result += "\\t"
-            default:
-                if scalar.value < 0x20 {
-                    result += String(format: "\\u%04X", scalar.value)
-                } else {
-                    result += String(scalar)
-                }
-            }
-        }
-        return result
-    }
-
-    func buildEmptyKeyResult(startTime: Date) -> PluginQueryResult {
-        PluginQueryResult(
-            columns: ["Key", "Type", "TTL", "Value"],
-            columnTypeNames: ["String", "RedisType", "RedisInt", "RedisRaw"],
-            rows: [],
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildStatusResult(_ message: String, startTime: Date) -> PluginQueryResult {
-        PluginQueryResult(
-            columns: ["status"],
-            columnTypeNames: ["String"],
-            rows: [[message].asCells],
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildGenericResult(_ result: RedisReply, startTime: Date) -> PluginQueryResult {
-        switch result {
-        case .string(let s), .status(let s):
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["String"],
-                rows: [[s].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .integer(let i):
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["Int64"],
-                rows: [[String(i)].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .data(let d):
-            let str = String(data: d, encoding: .utf8) ?? d.base64EncodedString()
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["String"],
-                rows: [[str].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .array(let items):
-            let rows = items.map { ([redisReplyToString($0)] as [String?]).asCells }
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["String"],
-                rows: rows,
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .error(let e):
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["String"],
-                rows: [[e].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-
-        case .null:
-            return PluginQueryResult(
-                columns: ["result"],
-                columnTypeNames: ["String"],
-                rows: [["(nil)"].asCells],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-    }
-
-    func redisReplyToString(_ reply: RedisReply) -> String {
-        switch reply {
-        case .string(let s), .status(let s), .error(let s): return s
-        case .integer(let i): return String(i)
-        case .data(let d): return String(data: d, encoding: .utf8) ?? d.base64EncodedString()
-        case .array(let items): return "[\(items.map { redisReplyToString($0) }.joined(separator: ", "))]"
-        case .null: return "(nil)"
-        }
-    }
-
-    func buildHashResult(_ result: RedisReply, startTime: Date) -> PluginQueryResult {
-        guard let items = result.arrayValue, !items.isEmpty else {
-            return PluginQueryResult(
-                columns: ["Field", "Value"],
-                columnTypeNames: ["String", "String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        var rows: [[PluginCellValue]] = []
-        var i = 0
-        while i + 1 < items.count {
-            rows.append([redisReplyToString(items[i]), redisReplyToString(items[i + 1])].asCells)
-            i += 2
-        }
-
-        return PluginQueryResult(
-            columns: ["Field", "Value"],
-            columnTypeNames: ["String", "String"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildListResult(_ result: RedisReply, startOffset: Int = 0, startTime: Date) -> PluginQueryResult {
-        guard let items = result.arrayValue else {
-            return PluginQueryResult(
-                columns: ["Index", "Value"],
-                columnTypeNames: ["Int64", "String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        let rows = items.enumerated().map { index, item -> [PluginCellValue] in
-            ([String(startOffset + index), redisReplyToString(item)] as [String?]).asCells
-        }
-
-        return PluginQueryResult(
-            columns: ["Index", "Value"],
-            columnTypeNames: ["Int64", "String"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildSetResult(_ result: RedisReply, startTime: Date) -> PluginQueryResult {
-        guard let items = result.arrayValue else {
-            return PluginQueryResult(
-                columns: ["Member"],
-                columnTypeNames: ["String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        let rows = items.map { ([redisReplyToString($0)] as [String?]).asCells }
-
-        return PluginQueryResult(
-            columns: ["Member"],
-            columnTypeNames: ["String"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildSortedSetResult(_ result: RedisReply, withScores: Bool, startTime: Date) -> PluginQueryResult {
-        guard let items = result.arrayValue else {
-            return PluginQueryResult(
-                columns: withScores ? ["Member", "Score"] : ["Member"],
-                columnTypeNames: withScores ? ["String", "Double"] : ["String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        if withScores {
-            var rows: [[PluginCellValue]] = []
-            var i = 0
-            while i + 1 < items.count {
-                rows.append([redisReplyToString(items[i]), redisReplyToString(items[i + 1])].asCells)
-                i += 2
-            }
-            return PluginQueryResult(
-                columns: ["Member", "Score"],
-                columnTypeNames: ["String", "Double"],
-                rows: rows,
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        } else {
-            let rows = items.map { ([redisReplyToString($0)] as [String?]).asCells }
-            return PluginQueryResult(
-                columns: ["Member"],
-                columnTypeNames: ["String"],
-                rows: rows,
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-    }
-
-    func buildStreamResult(_ result: RedisReply, startTime: Date) -> PluginQueryResult {
-        guard let entries = result.arrayValue else {
-            return PluginQueryResult(
-                columns: ["ID", "Fields"],
-                columnTypeNames: ["String", "String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        var rows: [[PluginCellValue]] = []
-        for entry in entries {
-            guard let entryParts = entry.arrayValue, entryParts.count >= 2,
-                  let fields = entryParts[1].arrayValue else {
-                continue
-            }
-            let entryId = redisReplyToString(entryParts[0])
-
-            var fieldPairs: [String] = []
-            var i = 0
-            while i + 1 < fields.count {
-                fieldPairs.append("\(redisReplyToString(fields[i]))=\(redisReplyToString(fields[i + 1]))")
-                i += 2
-            }
-            rows.append([entryId, fieldPairs.joined(separator: ", ")].asCells)
-        }
-
-        return PluginQueryResult(
-            columns: ["ID", "Fields"],
-            columnTypeNames: ["String", "String"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
-        )
-    }
-
-    func buildConfigResult(_ result: RedisReply, startTime: Date) -> PluginQueryResult {
-        guard let items = result.arrayValue, !items.isEmpty else {
-            return PluginQueryResult(
-                columns: ["Parameter", "Value"],
-                columnTypeNames: ["String", "String"],
-                rows: [],
-                rowsAffected: 0,
-                executionTime: Date().timeIntervalSince(startTime)
-            )
-        }
-
-        var rows: [[PluginCellValue]] = []
-        var i = 0
-        while i + 1 < items.count {
-            rows.append([redisReplyToString(items[i]), redisReplyToString(items[i + 1])].asCells)
-            i += 2
-        }
-
-        return PluginQueryResult(
-            columns: ["Parameter", "Value"],
-            columnTypeNames: ["String", "String"],
-            rows: rows,
-            rowsAffected: 0,
-            executionTime: Date().timeIntervalSince(startTime)
         )
     }
 }
